Modified repos & implemented login & registration

+++ java-server/src/main/java/com/eliasafara/javaserver/repository/ContactRepository.java
@@ -2,8 +2,6 @@ package com.eliasafara.javaserver.repository;
 
 import com.eliasafara.javaserver.datamodel.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface ContactRepository extends JpaRepository<Contact, Long> {
+public interface ContactRepository extends JpaRepository<Contact, Integer> {
 
 }
